enforce wait for all goroutines in client tests
client test code was opening connections in separate goroutines but not
waiting their termination. This is particularly dangerous if any calls
are made (asynchronously) to testing.*.Log* after the test function is
over. Not only the behavior is undefined, it was causing a data race in
go1.6 which is fixed in go1.7 by the following commit:

https://github.com/golang/go/commit/5c83e651adfa78d73634557cfaf2fbc9bde599f0